Move server code from gui namespace
