+++ sinner/server/api/ZMQClientAPI.py
@@ -6,10 +6,10 @@ from typing import Optional, Callable
 
 from zmq import Socket, ZMQError
 
-from sinner.gui.server.api.BaseClientAPI import BaseClientAPI
-from sinner.gui.server.api.messages.NotificationMessage import NotificationMessage
-from sinner.gui.server.api.messages.RequestMessage import RequestMessage
-from sinner.gui.server.api.messages.ResponseMessage import ResponseMessage
+from sinner.server.api.BaseClientAPI import BaseClientAPI
+from sinner.server.api.messages.NotificationMessage import NotificationMessage
+from sinner.server.api.messages.RequestMessage import RequestMessage
+from sinner.server.api.messages.ResponseMessage import ResponseMessage
 
 
 class ZMQClientAPI(BaseClientAPI):
