+++ sinner/server/api/messages/BaseMessage.py
